+++ src/components/react-hook-form/InputFormField.tsx
@@ -6,7 +6,7 @@ import {
   useFormContext,
 } from 'react-hook-form'
 import { InputField, InputFieldProps, Status } from '../form'
-import { useForwardedRef } from '../utils/useForwardedRef'
+import { useForwardedRef } from '../util/useForwardedRef'
 import { useFieldError } from './util/useFieldError'
 
 export type InputFormFieldProps<
